Update sharing post ui
