+++ src/components/sharing-post/DeleteSharingPost.graphql
@@ -1,0 +1,5 @@
+mutation DeleteSharingPost($id: ID!) {
+  deletePost(id: $id) {
+    id
+  }
+}
