+++ src/components/sharing-post/SharePost.graphql
