+++ src/hooks/useNeedToLogin.tsx
@@ -15,15 +15,22 @@ export default function useNeedToLogin(when = true) {
     if (!jwt && when) {
       toast.warn(
         <div>
-          로그인이 필요합니다.{' '}
-          <Link
-            href="/login"
-            onClick={() => sessionStorage.setItem('redirectToAfterLogin', router.asPath)}
-          >
-            로그인하기
-          </Link>
+          로그인이 필요합니다. <LoginLink />
         </div>
       )
     }
   }, [jwt, router, when])
+}
+
+export function LoginLink() {
+  const router = useRouter()
+
+  return (
+    <Link
+      href="/login"
+      onClick={() => sessionStorage.setItem('redirectToAfterLogin', router.asPath)}
+    >
+      로그인하기
+    </Link>
+  )
 }
